requirements.txt
bcrypt
flask
flask-mysqldb
flask-sqlqlchemy
flask-login
flask-bcrypt
dotenv
flask-wtf
flask-wtforms
